refactor: 공통 필드를 BaseEntity로 분리
