+++ 2-sprint-mission/src/main/java/com/sprint/mission/discodeit/entity/Channel.java
@@ -5,10 +5,7 @@ import java.util.HashSet;
 import java.util.Set;
 import java.util.UUID;
 
-public class Channel {
-    private final UUID id;
-    private final long createdAt;
-    private long updatedAt;
+public class Channel extends BaseEntity {
     private ChannelType type;
     private String category;
     private String name;
@@ -17,27 +14,13 @@ public class Channel {
     private UserRole writePermission;
 
     public Channel(ChannelType type, String category, String name, UUID userId, UserRole writePermission) {
-        this.id = UUID.randomUUID();
-        this.createdAt = System.currentTimeMillis();
-        this.updatedAt = this.createdAt;
+        super();
         this.type = type;
         this.category = category;
         this.name = name;
         this.userId = userId;
         this.userMembers = new HashSet<>();
         this.writePermission = writePermission;
-    }
-
-    public UUID getId() {
-        return id;
-    }
-
-    public long getCreatedAt() {
-        return createdAt;
-    }
-
-    public long getUpdatedAt() {
-        return updatedAt;
     }
 
     public ChannelType getType() {
@@ -67,40 +50,43 @@ public class Channel {
 
     public void updateType(ChannelType type) {
         this.type = type;
-        this.updatedAt = System.currentTimeMillis();
+        updateTimestamp();
     }
 
     public void updateCategory(String category) {
         this.category = category;
-        this.updatedAt = System.currentTimeMillis();
+        updateTimestamp();
     }
 
     public void updateName(String name) {
         this.name = name;
-        this.updatedAt = System.currentTimeMillis();
+        updateTimestamp();
     }
 
     public void updateWritePermission(UserRole writePermission) {
         this.writePermission = writePermission;
+        updateTimestamp();
     }
 
     public void addMember(UUID memberId){
         this.userMembers.add(memberId);
+        updateTimestamp();
     }
 
     public void removeMember(UUID memberId){
         this.userMembers.remove(memberId);
+        updateTimestamp();
     }
 
     @Override
     public String toString() {
         return "Channel{" +
-                "id=" + id +
-                ", type='" + type + '\'' +
+                "id=" + getId() +
+                ", type=" + type +
                 ", category='" + category + '\'' +
                 ", name='" + name + '\'' +
-                ", userId='" + userId + '\'' +
-                ", userMembers=" + userMembers + '\'' +
+                ", userId=" + userId +
+                ", userMembers=" + userMembers +
                 ", writePermission=" + writePermission +
                 '}';
     }
