+++ 2-sprint-mission/src/main/java/com/sprint/mission/discodeit/entity/User.java
@@ -1,11 +1,6 @@
 package com.sprint.mission.discodeit.entity;
 
-import java.util.UUID;
-
-public class User {
-    private final UUID id;
-    private final long createdAt;
-    private long updatedAt;
+public class User extends BaseEntity {
     private final String email;
     private String password;
     private String nickname;
@@ -13,26 +8,12 @@ public class User {
     private UserRole role;
 
     public User(String email, String password, String nickname, UserStatus status, UserRole role) {
-        this.id = UUID.randomUUID();
-        this.createdAt = System.currentTimeMillis();
-        this.updatedAt = this.createdAt;
+        super();
         this.email = email;
         this.password = password;
         this.nickname = nickname;
         this.status = status;
         this.role = role;
-    }
-
-    public UUID getId() {
-        return id;
-    }
-
-    public long getCreatedAt() {
-        return createdAt;
-    }
-
-    public long getUpdatedAt() {
-        return updatedAt;
     }
 
     public String getEmail() {
@@ -57,28 +38,28 @@ public class User {
 
     public void updatePassword(String password) {
         this.password = password;
-        this.updatedAt = System.currentTimeMillis();
+        updateTimestamp();
     }
 
     public void updateNickname(String nickname) {
         this.nickname = nickname;
-        this.updatedAt = System.currentTimeMillis();
+        updateTimestamp();
     }
 
     public void updateStatus(UserStatus status) {
         this.status = status;
-        this.updatedAt = System.currentTimeMillis();
+        updateTimestamp();
     }
 
     public void updateRole(UserRole role) {
         this.role = role;
-        this.updatedAt = System.currentTimeMillis();
+        updateTimestamp();
     }
 
     @Override
     public String toString() {
         return "User{" +
-                "id=" + id +
+                "id=" + getId() +
                 ", email='" + email + '\'' +
                 ", password='" + password + '\'' +
                 ", nickname='" + nickname + '\'' +
